requirements.txt
discord.py==2.3.2
requests==2.31.0
roblox==2.0.0
pytimeparse2==1.7.1
aiohttp~=3.8.1